SCons and CMake all tools
Refactor sdl/utils.cpp into two files. This allows building all tools without pulling in the entire engine.

Note that the size of some targets could be further reduced by additional refactoring in other areas.

SCons and CMake can now build all of the following targets:

    * campaignd
    * create_images
    * cutter
    * exploder
    * schema_generator
    * schema_validator
    * test
    * wesmage
    * wesnoth
    * wesnothd
