+++ src/CMakeLists.txt
@@ -368,6 +368,7 @@ set(wesnoth-sdl_SRC
 	sdl/rect.cpp
 	sdl/window.cpp
 	sdl/utils.cpp
+	sdl/utils_basic.cpp
 	xBRZ/xbrz.cpp
 )
 
@@ -1184,9 +1185,6 @@ set(exploder_SRC
 	tools/exploder_utils.cpp
 	tools/exploder_cutter.cpp
 	tools/exploder_composer.cpp
-	tools/dummy_video.cpp
-	sdl/utils.cpp
-	tracer.cpp
 )
 
 add_executable(exploder ${exploder_SRC})
@@ -1204,9 +1202,6 @@ set(cutter_SRC
 	tools/cutter.cpp
 	tools/exploder_utils.cpp
 	tools/exploder_cutter.cpp
-	tools/dummy_video.cpp
-	sdl/utils.cpp
-	tracer.cpp
 )
 
 add_executable(cutter ${cutter_SRC})
@@ -1221,7 +1216,6 @@ set_target_properties(cutter PROPERTIES OUTPUT_NAME ${BINARY_PREFIX}cutter${BINA
 install(TARGETS cutter DESTINATION ${BINDIR})
 
 set(schema_generator_SRC
-	tools/dummy_video.cpp
 	tools/schema/schema_generator.cpp
 	tools/schema/sourceparser.cpp
 	tools/schema/error_container.cpp
@@ -1245,7 +1239,6 @@ set(schema_validator_SRC
 	tools/validator/validator_tool.cpp
 	serialization/schema_validator.cpp
 	serialization/validator.cpp
-	tools/dummy_video.cpp
 	tools/schema/tag.cpp
 	filesystem_boost.cpp
 	filesystem_common.cpp
@@ -1271,10 +1264,7 @@ set(wesmage_SRC
 	wesmage/exit.cpp
 	wesmage/filter.cpp
 	wesmage/options.cpp
-	tools/dummy_video.cpp
 	tools/exploder_utils.cpp
-	sdl/utils.cpp
-	tracer.cpp
 )
 
 add_executable(wesmage ${wesmage_SRC})
@@ -1393,10 +1383,7 @@ if(ENABLE_TESTS)
 		# Due to its unique nature the program is never installed.
 		set(create_images_SRC
 			tests/create_images.cpp
-			tools/dummy_video.cpp
 			tools/exploder_utils.cpp
-			sdl/utils.cpp
-			tracer.cpp
 		)
 
 		add_executable(create_images ${create_images_SRC})
